Check for empty Groups Key
Right now if you don't set a Groups Key in the integration configuration then $this->getKey() will log "PHP Notice:  Undefined index:  " because it's being asked to look for a blank key. This commit first checks whether the key is set before passing it to getKey() to prevent this Notice.